Add files via upload
This is PHP code segment within an html document to access the information placed in the company_cars table in vehicle_fleet database on my localhost (mysql wampserver). This php code allows viewing of the information from the SQL table. I provided the user with three different ways of viewing the information. My html document allow the user to choose a specific output. These choices allow a user to view the information sorted in ascending or descending order, according to Make, Model, or Mileage.